nvme-print: fix integer overflow in EOM descriptor offset
json_phy_rx_eom_descs() computed the vendor-specific eye data offset
as a plain int multiply of two device-controlled 16-bit values, which
can overflow and sign-extend into a huge offset once added to the
descriptor size. stdout_phy_rx_eom_descs() already avoids this with a
wider cast; apply the same cast here.

+++ src/nvme-print-json.c
@@ -2382,7 +2382,7 @@ static void json_phy_rx_eom_descs(struct nvme_phy_rx_eom_log *log,
 		}
 
 		/* Hex dump Vendor Specific Eye Data */
-		vsdataoffset = (nrows * ncols) + sizeof(struct nvme_eom_lane_desc);
+		vsdataoffset = (unsigned int)nrows * ncols + sizeof(struct nvme_eom_lane_desc);
 		vsdata = (unsigned char *)((unsigned char *)desc + vsdataoffset);
 
 		hexdata = hexstr;
